requirements.txt
asgiref==3.5.2
black==22.6.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
psycopg2==2.9.3
drf-spectacular==0.23.1
flake8==5.0.3
